4137 fix c++ s3 file reference in c++ sdk guide (#4243)
* moved get/put objectacl into one folder to match reference in SDK guide.
renamed gtests to tests

* Fixes that somehow broke

* fixing yaml
