+++ cpp/example_code/s3/README.md
@@ -40,7 +40,7 @@ Code excerpts that show you how to call individual service functions.
 - [Get the access control list (ACL) for an S3 bucket](./get_bucket_acl.cpp) (GetBucketAcl)
 - [Get a bucket policy (permission to access resources) for an S3 bucket](./get_bucket_policy.cpp) (GetBucketPolicy)
 - [Get an object out of an S3 bucket](./get_object.cpp) (GetObject)
-- [Get the access control list (ACL) for an object in an S3 bucket](./get_object_acl.cpp) (GetObjectAcl)
+- [Get the access control list (ACL) for an object in an S3 bucket](./get_put_object_acl.cpp) (GetObjectAcl)
 - [Get and set the access control list (ACL) for an S3 bucket](./get_put_bucket_acl.cpp) (GetBucketAcl, PutBucketAcl)
 - [Get the configuration of an S3 bucket configured for static website hosting](./get_website_config.cpp) (
   GetBucketWebsite)
@@ -51,7 +51,7 @@ Code excerpts that show you how to call individual service functions.
 - [Set the access control list (ACL) for an S3 bucket](./put_bucket_acl.cpp) (PutBucketAcl)
 - [Add a bucket policy (permission to access resources) to an S3 bucket](./put_bucket_policy.cpp) (PutBucketPolicy)
 - [Upload an object to an S3 bucket](./put_object.cpp) (PutObject)
-- [Set the access control list (ACL) for an object in an S3 bucket](./put_object_acl.cpp) (PutObjectAcl)
+- [Set the access control list (ACL) for an object in an S3 bucket](./get_put_object_acl.cpp) (PutObjectAcl)
 - [Upload an object to an S3 bucket (asynchronously)](./put_object_async.cpp) (PutObjectAsync)
 - [Upload an object to an S3 bucket (using a memory buffer instead of local disk copy)](./put_object_buffer.cpp) (
   PutObject)
