--- cpp/example_code/s3/get_object_acl.cpp
@@ -1,187 +1,0 @@
-/*
-   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-   SPDX-License-Identifier: Apache-2.0
-*/
-
-#include <iostream>
-#include <aws/core/Aws.h>
-#include <aws/s3/S3Client.h>
-#include <aws/s3/model/GetBucketAclRequest.h>
-#include <aws/s3/model/GetObjectAclRequest.h>
-#include <aws/s3/model/Grant.h>
-#include <aws/s3/model/Permission.h>
-#include "awsdoc/s3/s3_examples.h"
-
-/**
- * Before running this C++ code example, set up your development environment, including your credentials.
- *
- * For more information, see the following documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-cpp/v1/developer-guide/getting-started.html
- *
- * Purpose
- *
- * Demonstrates using the AWS SDK for C++ to get the access control list (ACL) for an object
- * in an S3 bucket.
- *
- */
-
-static Aws::String GetGranteeTypeString(const Aws::S3::Model::Type &type);
-
-static Aws::String GetPermissionString(const Aws::S3::Model::Permission &permission);
-
-//! Routine which demonstrates getting the ACL for an object in an S3 bucket.
-/*!
-  \fn GetObjectAcl()
-  \param bucketName Name of the bucket.
-  \param objectKey Name of the object in the bucket.
-  \param clientConfig Aws client configuration.
-*/
-
-// snippet-start:[s3.cpp.get_object_acl.code]
-bool AwsDoc::S3::GetObjectAcl(const Aws::String &bucketName,
-                              const Aws::String &objectKey,
-                              const Aws::Client::ClientConfiguration &clientConfig) {
-    Aws::S3::S3Client s3_client(clientConfig);
-
-    Aws::S3::Model::GetObjectAclRequest request;
-    request.SetBucket(bucketName);
-    request.SetKey(objectKey);
-
-    Aws::S3::Model::GetObjectAclOutcome outcome =
-            s3_client.GetObjectAcl(request);
-
-    if (!outcome.IsSuccess()) {
-        const Aws::S3::S3Error &err = outcome.GetError();
-        std::cerr << "Error: GetObjectAcl: "
-                  << err.GetExceptionName() << ": " << err.GetMessage() << std::endl;
-    }
-    else {
-        Aws::Vector<Aws::S3::Model::Grant> grants =
-                outcome.GetResult().GetGrants();
-
-        for (auto it = grants.begin(); it != grants.end(); it++) {
-            std::cout << "For object " << objectKey << ": "
-                      << std::endl << std::endl;
-
-            Aws::S3::Model::Grant grant = *it;
-            Aws::S3::Model::Grantee grantee = grant.GetGrantee();
-
-            if (grantee.TypeHasBeenSet()) {
-                std::cout << "Type:          "
-                          << GetGranteeTypeString(grantee.GetType()) << std::endl;
-            }
-
-            if (grantee.DisplayNameHasBeenSet()) {
-                std::cout << "Display name:  "
-                          << grantee.GetDisplayName() << std::endl;
-            }
-
-            if (grantee.EmailAddressHasBeenSet()) {
-                std::cout << "Email address: "
-                          << grantee.GetEmailAddress() << std::endl;
-            }
-
-            if (grantee.IDHasBeenSet()) {
-                std::cout << "ID:            "
-                          << grantee.GetID() << std::endl;
-            }
-
-            if (grantee.URIHasBeenSet()) {
-                std::cout << "URI:           "
-                          << grantee.GetURI() << std::endl;
-            }
-
-            std::cout << "Permission:    " <<
-                      GetPermissionString(grant.GetPermission()) <<
-                      std::endl << std::endl;
-        }
-    }
-
-    return outcome.IsSuccess();
-}
-
-//! Routine which converts a built-in type enumeration to a human-readable string.
-/*!
- \fn GetGranteeTypeString()
- \param type Type enumeration.
-*/
-
-Aws::String GetGranteeTypeString(const Aws::S3::Model::Type &type) {
-    switch (type) {
-        case Aws::S3::Model::Type::AmazonCustomerByEmail:
-            return "Email address of an AWS account";
-        case Aws::S3::Model::Type::CanonicalUser:
-            return "Canonical user ID of an AWS account";
-        case Aws::S3::Model::Type::Group:
-            return "Predefined Amazon S3 group";
-        case Aws::S3::Model::Type::NOT_SET:
-            return "Not set";
-        default:
-            return "Type unknown";
-    }
-}
-
-//! Routine which converts a built-in type enumeration to a human-readable string.
-/*!
- \fn GetPermissionString()
- \param permission Permission enumeration.
-*/
-
-Aws::String GetPermissionString(const Aws::S3::Model::Permission &permission) {
-    switch (permission) {
-        case Aws::S3::Model::Permission::FULL_CONTROL:
-            return "Can read this object's data and its metadata, "
-                   "and read/write this object's permissions";
-        case Aws::S3::Model::Permission::NOT_SET:
-            return "Permission not set";
-        case Aws::S3::Model::Permission::READ:
-            return "Can read this object's data and its metadata";
-        case Aws::S3::Model::Permission::READ_ACP:
-            return "Can read this object's permissions";
-            // case Aws::S3::Model::Permission::WRITE // Not applicable.
-        case Aws::S3::Model::Permission::WRITE_ACP:
-            return "Can write this object's permissions";
-        default:
-            return "Permission unknown";
-    }
-}
-
-// snippet-end:[s3.cpp.get_object_acl.code]
-
-/*
- *
- * main function
- *
- * Prerequisites: Create an S3 bucket and object to get the ACL information about them.
- *
- * TODO(user): items: Set the following variables
- * - bucketName: The name of the bucket to get the ACL information about.
- * - object_name: Name of object already in bucket.
- *
- */
-
-#ifndef TESTING_BUILD
-
-int main() {
-    //TODO(user): Name of your bucket that already contains an object.
-    //See create_bucket.cpp and put_object.cpp to create a bucket and load an object into that bucket.
-    Aws::String bucket_name = "<Enter a bucket name>";
-    //TODO(user): Name of object already in bucket.
-    Aws::String object_name = "<Enter object name>";
-
-    Aws::SDKOptions options;
-    Aws::InitAPI(options);
-    {
-        Aws::Client::ClientConfiguration clientConfig;
-        // Optional: Set to the AWS Region in which the bucket was created (overrides config file).
-        // clientConfig.region = "us-east-1";
-        AwsDoc::S3::GetObjectAcl(bucket_name, object_name, clientConfig);
-    }
-    Aws::ShutdownAPI(options);
-
-    return 0;
-}
-
-#endif  // TESTING_BUILD
-
